Implement master 'layer layer's
Added the ability to switch layers using homerow layer selection available on a
momentary layer accessed by tap-dance thumbkeys. The choice of thumbkey
determines the nature of layer switching - currently supports layer holding
(only the thumbkey needs to be held), toggling (the master layer is cleared
after a layer is toggled), oneshot layer access, and oneshot with shift active
(could be extended to other mods easily using the same logic). This all works
nicely in my use-case/with my keymap, but I've not done extensive testing so
it could be brittle under other circumstances, or break parts of QMK that I
don't use. It's a relatively simple extension to tap-dance but it took a bit
of trial-and-error to get good behaviour, so maybe this can save someone else
from doing the same work again. Particularly useful for small keyboards and/or
many layers.